v1.0.0

+++ script/unprep.sh
@@ -2,13 +2,7 @@
 
 set -e
 
-DIR=$1
-if [ "${DIR}" = "" ]; then
-	DIR=`pwd`
-fi
-
 [ -d /usr/src/sys ] || exit 1
-[ -d ${DIR} ] || exit 1
 
 HAMMER2_LNK=/usr/src/sys/hammer2
 if [ ! -e ${HAMMER2_LNK} ]; then
@@ -16,6 +10,13 @@ if [ ! -e ${HAMMER2_LNK} ]; then
 	exit 1
 fi
 
+ICRC32_LNK=/usr/src/sys/lib/libkern/icrc32.c
+if [ ! -e ${ICRC32_LNK} ]; then
+	echo "${ICRC32_LNK} does not exist"
+	exit 1
+fi
+
 /bin/rm ${HAMMER2_LNK}
+/bin/rm ${ICRC32_LNK}
 
 echo "unprep success"
